Add  new ips file.
hey michenriksen, I think it would be cool too if the tool generated a ips.txt file. Well, sometimes I have to parse the hosts file and then move on to nmap to do a scan with some scripts. If it gets bad, what did I do you could give me an answer? I'm studying ruby ​​and I want to contribute to tool